Removed lombok, added getters

+++ src/main/java/org/payments/payments/PaymentNotFoundException.java
@@ -1,7 +1,7 @@
 package org.payments.payments;
 
 public class PaymentNotFoundException extends RuntimeException {
-    public PaymentNotFoundException(Long id) {
+    public PaymentNotFoundException(Integer id) {
         super ("Could not find payment " + id);
     }
 }
